update ui fp pro inv docs
update ui fp pro inv docs

+++ src/modules/production/finishing-printing/fp-retur-pro-inv-docs/view.js
@@ -1,14 +1,17 @@
 import { inject, Lazy } from 'aurelia-framework';
 import { Router } from 'aurelia-router';
 import { Service } from './service';
+import { Dialog } from '../../../../components/dialog/dialog';
+import { AlertView } from './custom-dialog-view/alert-view';
 var moment = require("moment");
 
-@inject(Router, Service)
+@inject(Router, Service, Dialog)
 export class View {
 
-    constructor(router, service) {
+    constructor(router, service, dialog) {
         this.router = router;
         this.service = service;
+        this.dialog = dialog;
     }
 
     async activate(params) {
@@ -21,16 +24,16 @@ export class View {
         this.router.navigateToRoute('list');
     }
 
-
     deleteCallback(event) {
-        this.service.delete(this.data).then(result => {
-            alert(`delete data success`);
-            this.cancelCallback();
-        });
-    }
-
-    editCallback(event) {
-        this.router.navigateToRoute('edit', { id: this.data.Id });
+        this.dialog.show(AlertView, { message: "<div>Apakah anda yakin akan menghapus data ini?</div>" })
+            .then(response => {
+                if (!response.wasCancelled) {
+                    this.service.delete(this.data).then(result => {
+                        alert(`delete data success`);
+                        this.cancelCallback();
+                    });
+                }
+            });
     }
 
 }
